Create model interface
